data-processing/requirements.txt
supabase
python-dotenv
pandas
sqlalchemy
psycopg2